// src/app/interfaces/grupos.ts
export interface Grupo {
    id_grupo?: number;
    nombre_compania?: string;
    nombre_grupo?: string;
    cantidad_vehiculo?: number;
    status?: boolean;
    logo?: string;
    email?: string;
    pais?: string;
    ciudad?: string;
    direccion_1?: string;
    direccion_2?: string;
    fecha_creado?: string;
    fecha_actualizado?: string;
}